Makefile: add support for XIP and RAM runtime model
This adds support to either use the XIP (from FLASH) or RAM runtime
model for the slboot application.

Default is RAM runtime model.

This also includes some rework on the ARGFLAGS basically reoredering
and dropping some unneded flags.

And consolidate the STACK and HEAP size. The default STACK size will
be 0x200 (512B) and the HEAP size will be 0.

Signed-off-by: Alexander Herfurtner <mission@alexanderherfurtner.de>